+++ modules/sde-core/src/main/java/org/eclipse/tractusx/sde/core/processreport/model/ConsumerDownloadHistory.java
@@ -1,0 +1,49 @@
+/********************************************************************************
+ * Copyright (c) 2024 T-Systems International GmbH
+ * Copyright (c) 2024 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ********************************************************************************/
+package org.eclipse.tractusx.sde.core.processreport.model;
+
+import java.time.LocalDateTime;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class ConsumerDownloadHistory {
+	
+
+	private String processId;
+	private String providerUrl;
+	private String connectorId;
+	private Integer numberOfItems;
+	private Integer downloadFailed;
+	private Integer downloadSuccessed;
+	private String status;
+	private LocalDateTime startDate;
+	private LocalDateTime endDate;
+	private Object offers;
+	private Object policies;
+    private String referenceProcessId;
+
+}